Refactor Zsh installation scripts to include checks for existing installations and improve alias setup in .zshrc

+++ install_zsh_1.sh
@@ -2,6 +2,15 @@
 set -e
 
 sudo apt update && sudo apt upgrade -y
-sudo apt install zsh -y
-sudo apt-get install fonts-powerline -y
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+
+if ! command -v zsh &> /dev/null; then
+    sudo apt install zsh -y
+fi
+
+if ! dpkg -s fonts-powerline &> /dev/null; then
+    sudo apt-get install fonts-powerline -y
+fi
+
+if [ ! -d "$HOME/.oh-my-zsh" ]; then
+    sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+fi
